+++ src/backend/macos/file_dialog.rs
@@ -1,13 +1,14 @@
-mod panel_ffi;
-use panel_ffi::Panel;
-
-use crate::backend::DialogFutureType;
-use crate::{FileDialog, FileHandle};
-
+use objc2::rc::autoreleasepool;
+use objc2_app_kit::NSModalResponseOK;
 use std::path::PathBuf;
 
+mod panel_ffi;
+
+use self::panel_ffi::Panel;
 use super::modal_future::ModalFuture;
-use super::utils::{run_on_main, INSWindow, NSWindow};
+use super::utils::{run_on_main, window_from_raw_window_handle};
+use crate::backend::DialogFutureType;
+use crate::{FileDialog, FileHandle};
 
 //
 // File Picker
@@ -16,11 +17,11 @@ use super::utils::{run_on_main, INSWindow, NSWindow};
 use crate::backend::FilePickerDialogImpl;
 impl FilePickerDialogImpl for FileDialog {
     fn pick_file(self) -> Option<PathBuf> {
-        objc::rc::autoreleasepool(move || {
-            run_on_main(move || {
-                let panel = Panel::build_pick_file(&self);
+        autoreleasepool(move |_| {
+            run_on_main(move |mtm| {
+                let panel = Panel::build_pick_file(&self, mtm);
 
-                if panel.run_modal() == 1 {
+                if panel.run_modal() == NSModalResponseOK {
                     Some(panel.get_result())
                 } else {
                     None
@@ -30,11 +31,11 @@ impl FilePickerDialogImpl for FileDialog {
     }
 
     fn pick_files(self) -> Option<Vec<PathBuf>> {
-        objc::rc::autoreleasepool(move || {
-            run_on_main(move || {
-                let panel = Panel::build_pick_files(&self);
+        autoreleasepool(move |_| {
+            run_on_main(move |mtm| {
+                let panel = Panel::build_pick_files(&self, mtm);
 
-                if panel.run_modal() == 1 {
+                if panel.run_modal() == NSModalResponseOK {
                     Some(panel.get_results())
                 } else {
                     None
@@ -47,13 +48,13 @@ impl FilePickerDialogImpl for FileDialog {
 use crate::backend::AsyncFilePickerDialogImpl;
 impl AsyncFilePickerDialogImpl for FileDialog {
     fn pick_file_async(self) -> DialogFutureType<Option<FileHandle>> {
-        let win = self.parent.as_ref().map(NSWindow::from_raw_window_handle);
+        let win = self.parent.as_ref().map(window_from_raw_window_handle);
 
         let future = ModalFuture::new(
             win,
-            move || Panel::build_pick_file(&self),
+            move |mtm| Panel::build_pick_file(&self, mtm),
             |panel, res_id| {
-                if res_id == 1 {
+                if res_id == NSModalResponseOK {
                     Some(panel.get_result().into())
                 } else {
                     None
@@ -65,13 +66,13 @@ impl AsyncFilePickerDialogImpl for FileDialog {
     }
 
     fn pick_files_async(self) -> DialogFutureType<Option<Vec<FileHandle>>> {
-        let win = self.parent.as_ref().map(NSWindow::from_raw_window_handle);
+        let win = self.parent.as_ref().map(window_from_raw_window_handle);
 
         let future = ModalFuture::new(
             win,
-            move || Panel::build_pick_files(&self),
+            move |mtm| Panel::build_pick_files(&self, mtm),
             |panel, res_id| {
-                if res_id == 1 {
+                if res_id == NSModalResponseOK {
                     Some(
                         panel
                             .get_results()
@@ -96,10 +97,10 @@ impl AsyncFilePickerDialogImpl for FileDialog {
 use crate::backend::FolderPickerDialogImpl;
 impl FolderPickerDialogImpl for FileDialog {
     fn pick_folder(self) -> Option<PathBuf> {
-        objc::rc::autoreleasepool(move || {
-            run_on_main(move || {
-                let panel = Panel::build_pick_folder(&self);
-                if panel.run_modal() == 1 {
+        autoreleasepool(move |_| {
+            run_on_main(move |mtm| {
+                let panel = Panel::build_pick_folder(&self, mtm);
+                if panel.run_modal() == NSModalResponseOK {
                     Some(panel.get_result())
                 } else {
                     None
@@ -109,10 +110,10 @@ impl FolderPickerDialogImpl for FileDialog {
     }
 
     fn pick_folders(self) -> Option<Vec<PathBuf>> {
-        objc::rc::autoreleasepool(move || {
-            run_on_main(move || {
-                let panel = Panel::build_pick_folders(&self);
-                if panel.run_modal() == 1 {
+        autoreleasepool(move |_| {
+            run_on_main(move |mtm| {
+                let panel = Panel::build_pick_folders(&self, mtm);
+                if panel.run_modal() == NSModalResponseOK {
                     Some(panel.get_results())
                 } else {
                     None
@@ -125,13 +126,13 @@ impl FolderPickerDialogImpl for FileDialog {
 use crate::backend::AsyncFolderPickerDialogImpl;
 impl AsyncFolderPickerDialogImpl for FileDialog {
     fn pick_folder_async(self) -> DialogFutureType<Option<FileHandle>> {
-        let win = self.parent.as_ref().map(NSWindow::from_raw_window_handle);
+        let win = self.parent.as_ref().map(window_from_raw_window_handle);
 
         let future = ModalFuture::new(
             win,
-            move || Panel::build_pick_folder(&self),
+            move |mtm| Panel::build_pick_folder(&self, mtm),
             |panel, res_id| {
-                if res_id == 1 {
+                if res_id == NSModalResponseOK {
                     Some(panel.get_result().into())
                 } else {
                     None
@@ -143,13 +144,13 @@ impl AsyncFolderPickerDialogImpl for FileDialog {
     }
 
     fn pick_folders_async(self) -> DialogFutureType<Option<Vec<FileHandle>>> {
-        let win = self.parent.as_ref().map(NSWindow::from_raw_window_handle);
+        let win = self.parent.as_ref().map(window_from_raw_window_handle);
 
         let future = ModalFuture::new(
             win,
-            move || Panel::build_pick_folders(&self),
+            move |mtm| Panel::build_pick_folders(&self, mtm),
             |panel, res_id| {
-                if res_id == 1 {
+                if res_id == NSModalResponseOK {
                     Some(
                         panel
                             .get_results()
@@ -174,10 +175,10 @@ impl AsyncFolderPickerDialogImpl for FileDialog {
 use crate::backend::FileSaveDialogImpl;
 impl FileSaveDialogImpl for FileDialog {
     fn save_file(self) -> Option<PathBuf> {
-        objc::rc::autoreleasepool(move || {
-            run_on_main(move || {
-                let panel = Panel::build_save_file(&self);
-                if panel.run_modal() == 1 {
+        autoreleasepool(move |_| {
+            run_on_main(move |mtm| {
+                let panel = Panel::build_save_file(&self, mtm);
+                if panel.run_modal() == NSModalResponseOK {
                     Some(panel.get_result())
                 } else {
                     None
@@ -190,13 +191,13 @@ impl FileSaveDialogImpl for FileDialog {
 use crate::backend::AsyncFileSaveDialogImpl;
 impl AsyncFileSaveDialogImpl for FileDialog {
     fn save_file_async(self) -> DialogFutureType<Option<FileHandle>> {
-        let win = self.parent.as_ref().map(NSWindow::from_raw_window_handle);
+        let win = self.parent.as_ref().map(window_from_raw_window_handle);
 
         let future = ModalFuture::new(
             win,
-            move || Panel::build_save_file(&self),
+            move |mtm| Panel::build_save_file(&self, mtm),
             |panel, res_id| {
-                if res_id == 1 {
+                if res_id == NSModalResponseOK {
                     Some(panel.get_result().into())
                 } else {
                     None