+++ src/backend/macos/file_dialog/panel_ffi.rs
@@ -1,115 +1,133 @@
-use crate::FileDialog;
-
-use std::mem;
+use core::str;
 use std::path::Path;
-use std::{ops::DerefMut, path::PathBuf};
+use std::path::PathBuf;
 
-use objc::{class, msg_send, sel, sel_impl};
-use objc_id::{Id, Shared};
+use block2::Block;
+use objc2::rc::Id;
 
-use super::super::utils::{INSURL, NSURL};
-
-use crate::backend::macos::utils::{INSWindow, NSWindow};
-use objc::runtime::{Object, YES};
-use objc::runtime::{BOOL, NO};
-use objc_foundation::{INSArray, INSString, NSArray, NSString};
+use objc2_app_kit::{NSModalResponse, NSOpenPanel, NSSavePanel, NSWindow, NSWindowLevel};
+use objc2_foundation::{MainThreadMarker, NSArray, NSObjectProtocol, NSString, NSURL};
 use raw_window_handle::RawWindowHandle;
 
 use super::super::{
+    modal_future::{AsModal, InnerModal},
     utils::{FocusManager, PolicyManager},
-    AsModal,
 };
+use crate::backend::macos::utils::window_from_raw_window_handle;
+use crate::FileDialog;
 
 extern "C" {
     pub fn CGShieldingWindowLevel() -> i32;
 }
 
-fn make_nsstring(s: &str) -> Id<NSString> {
-    NSString::from_str(s)
-}
-
 pub struct Panel {
-    pub(crate) panel: Id<Object>,
-    parent: Option<Id<NSWindow, Shared>>,
+    // Either `NSSavePanel` or the subclass `NSOpenPanel`
+    pub(crate) panel: Id<NSSavePanel>,
+    parent: Option<Id<NSWindow>>,
     _focus_manager: FocusManager,
     _policy_manager: PolicyManager,
 }
 
 impl AsModal for Panel {
-    fn modal_ptr(&mut self) -> *mut Object {
-        self.panel.deref_mut()
+    fn inner_modal(&self) -> &NSSavePanel {
+        &*self.panel
+    }
+}
+
+impl InnerModal for NSSavePanel {
+    fn begin_modal(&self, window: &NSWindow, handler: &Block<dyn Fn(NSModalResponse)>) {
+        unsafe { self.beginSheetModalForWindow_completionHandler(window, handler) }
+    }
+
+    fn run_modal(&self) -> NSModalResponse {
+        unsafe { self.runModal() }
     }
 }
 
 impl Panel {
-    pub fn new(panel: *mut Object) -> Self {
-        let _policy_manager = PolicyManager::new();
+    fn as_open_panel(&self) -> Option<&NSOpenPanel> {
+        if self.panel.is_kind_of::<NSOpenPanel>() {
+            let ptr: *const NSSavePanel = &*self.panel;
+            let ptr: *const NSOpenPanel = ptr.cast();
+            // SAFETY: Just checked that the panel is an `NSOpenPanel`
+            Some(unsafe { &*ptr })
+        } else {
+            None
+        }
+    }
 
-        let _focus_manager = FocusManager::new();
+    pub fn new(panel: Id<NSSavePanel>, parent: Option<&RawWindowHandle>) -> Self {
+        let _policy_manager = PolicyManager::new(MainThreadMarker::from(&*panel));
 
-        let _: () = unsafe { msg_send![panel, setLevel: CGShieldingWindowLevel()] };
+        let _focus_manager = FocusManager::new(MainThreadMarker::from(&*panel));
+
+        panel.setLevel(unsafe { CGShieldingWindowLevel() } as NSWindowLevel);
         Self {
-            panel: unsafe { Id::from_ptr(panel) },
+            panel,
+            parent: parent.map(window_from_raw_window_handle),
             _focus_manager,
             _policy_manager,
-            parent: None,
         }
     }
 
-    pub fn open_panel() -> Self {
-        Self::new(unsafe { msg_send![class!(NSOpenPanel), openPanel] })
-    }
-
-    pub fn save_panel() -> Self {
-        Self::new(unsafe { msg_send![class!(NSSavePanel), savePanel] })
-    }
-
-    pub fn run_modal(&self) -> i32 {
+    pub fn run_modal(&self) -> NSModalResponse {
         if let Some(parent) = self.parent.clone() {
-            let completion = { block::ConcreteBlock::new(|_: isize| {}) };
+            let completion = block2::StackBlock::new(|_: isize| {});
 
             unsafe {
-                msg_send![self.panel, beginSheetModalForWindow: parent completionHandler: &completion]
+                self.panel
+                    .beginSheetModalForWindow_completionHandler(&parent, &completion)
             }
-
-            mem::forget(completion);
         }
 
-        unsafe { msg_send![self.panel, runModal] }
+        unsafe { self.panel.runModal() }
     }
 
-    pub fn set_can_choose_directories(&self, v: BOOL) {
-        let _: () = unsafe { msg_send![self.panel, setCanChooseDirectories: v] };
+    pub fn get_result(&self) -> PathBuf {
+        unsafe {
+            let url = self.panel.URL().unwrap();
+            url.path().unwrap().to_string().into()
+        }
     }
 
-    pub fn set_can_create_directories(&self, v: BOOL) {
-        let _: () = unsafe { msg_send![self.panel, setCanCreateDirectories: v] };
+    pub fn get_results(&self) -> Vec<PathBuf> {
+        unsafe {
+            let urls = self.as_open_panel().unwrap().URLs();
+
+            let mut res = Vec::new();
+            for url in urls {
+                res.push(url.path().unwrap().to_string().into());
+            }
+
+            res
+        }
+    }
+}
+
+trait PanelExt {
+    fn panel(&self) -> &NSSavePanel;
+
+    fn set_can_create_directories(&self, can: bool) {
+        unsafe { self.panel().setCanCreateDirectories(can) }
     }
 
-    pub fn set_can_choose_files(&self, v: BOOL) {
-        let _: () = unsafe { msg_send![self.panel, setCanChooseFiles: v] };
-    }
-
-    pub fn set_allows_multiple_selection(&self, v: BOOL) {
-        let _: () = unsafe { msg_send![self.panel, setAllowsMultipleSelection: v] };
-    }
-
-    pub fn add_filters(&self, params: &FileDialog) {
+    fn add_filters(&self, opt: &FileDialog) {
         let mut exts: Vec<String> = Vec::new();
 
-        for filter in params.filters.iter() {
+        for filter in opt.filters.iter() {
             exts.append(&mut filter.extensions.to_vec());
         }
 
-        unsafe {
-            let f_raw: Vec<_> = exts.iter().map(|ext| make_nsstring(&ext)).collect();
-            let array = NSArray::from_vec(f_raw);
+        let f_raw: Vec<_> = exts.iter().map(|ext| NSString::from_str(&ext)).collect();
+        let array = NSArray::from_vec(f_raw);
 
-            let _: () = msg_send![self.panel, setAllowedFileTypes: array];
+        unsafe {
+            #[allow(deprecated)]
+            self.panel().setAllowedFileTypes(Some(&array));
         }
     }
 
-    pub fn set_path(&self, path: &Path, file_name: Option<&str>) {
+    fn set_path(&self, path: &Path, file_name: Option<&str>) {
         // if file_name is some, and path is a dir
         let path = if let (Some(name), true) = (file_name, path.is_dir()) {
             let mut path = path.to_owned();
@@ -122,56 +140,39 @@ impl Panel {
 
         if let Some(path) = path.to_str() {
             unsafe {
-                let url = NSURL::file_url_with_path(path, true);
-                let () = msg_send![self.panel, setDirectoryURL: url];
+                let url = NSURL::fileURLWithPath_isDirectory(&NSString::from_str(path), true);
+                self.panel().setDirectoryURL(Some(&url));
             }
         }
     }
 
-    pub fn set_file_name(&self, file_name: &str) {
+    fn set_file_name(&self, file_name: &str) {
         unsafe {
-            let file_name = make_nsstring(file_name);
-            let () = msg_send![self.panel, setNameFieldStringValue: file_name];
+            self.panel()
+                .setNameFieldStringValue(&NSString::from_str(file_name))
         }
     }
 
-    pub fn set_title(&self, title: &str) {
-        unsafe {
-            let title = make_nsstring(title);
-            let () = msg_send![self.panel, setMessage: title];
-        }
+    fn set_title(&self, title: &str) {
+        unsafe { self.panel().setMessage(Some(&NSString::from_str(title))) }
     }
+}
 
-    pub fn set_parent(&mut self, parent: &RawWindowHandle) {
-        self.parent = Some(NSWindow::from_raw_window_handle(parent).share());
+impl PanelExt for Id<NSSavePanel> {
+    fn panel(&self) -> &NSSavePanel {
+        &self
     }
+}
 
-    pub fn get_result(&self) -> PathBuf {
-        unsafe {
-            let url = msg_send![self.panel, URL];
-            let url: Id<NSURL> = Id::from_ptr(url);
-            url.to_path_buf()
-        }
-    }
-
-    pub fn get_results(&self) -> Vec<PathBuf> {
-        unsafe {
-            let urls = msg_send![self.panel, URLs];
-            let urls: Id<NSArray<NSURL>> = Id::from_ptr(urls);
-
-            let mut res = Vec::new();
-            for url in urls.to_vec() {
-                res.push(url.to_path_buf());
-            }
-
-            res
-        }
+impl PanelExt for Id<NSOpenPanel> {
+    fn panel(&self) -> &NSSavePanel {
+        &self
     }
 }
 
 impl Panel {
-    pub fn build_pick_file(opt: &FileDialog) -> Self {
-        let mut panel = Panel::open_panel();
+    pub fn build_pick_file(opt: &FileDialog, mtm: MainThreadMarker) -> Self {
+        let panel = unsafe { NSOpenPanel::openPanel(mtm) };
 
         if !opt.filters.is_empty() {
             panel.add_filters(&opt);
@@ -189,25 +190,18 @@ impl Panel {
             panel.set_title(title);
         }
 
-        if let Some(parent) = &opt.parent {
-            panel.set_parent(parent);
-        }
-
         if let Some(can) = opt.can_create_directories {
-            panel.set_can_create_directories(match can {
-                true => YES,
-                false => NO,
-            });
+            panel.set_can_create_directories(can);
         }
 
-        panel.set_can_choose_directories(NO);
-        panel.set_can_choose_files(YES);
+        unsafe { panel.setCanChooseDirectories(false) };
+        unsafe { panel.setCanChooseFiles(true) };
 
-        panel
+        Self::new(Id::into_super(panel), opt.parent.as_ref())
     }
 
-    pub fn build_save_file(opt: &FileDialog) -> Self {
-        let mut panel = Panel::save_panel();
+    pub fn build_save_file(opt: &FileDialog, mtm: MainThreadMarker) -> Self {
+        let panel = unsafe { NSSavePanel::savePanel(mtm) };
 
         if !opt.filters.is_empty() {
             panel.add_filters(&opt);
@@ -225,22 +219,15 @@ impl Panel {
             panel.set_title(title);
         }
 
-        if let Some(parent) = &opt.parent {
-            panel.set_parent(parent);
-        }
-
         if let Some(can) = opt.can_create_directories {
-            panel.set_can_create_directories(match can {
-                true => YES,
-                false => NO,
-            });
+            panel.set_can_create_directories(can);
         }
 
-        panel
+        Self::new(panel, opt.parent.as_ref())
     }
 
-    pub fn build_pick_folder(opt: &FileDialog) -> Self {
-        let mut panel = Panel::open_panel();
+    pub fn build_pick_folder(opt: &FileDialog, mtm: MainThreadMarker) -> Self {
+        let panel = unsafe { NSOpenPanel::openPanel(mtm) };
 
         if let Some(path) = &opt.starting_directory {
             panel.set_path(path, opt.file_name.as_deref());
@@ -248,54 +235,40 @@ impl Panel {
 
         if let Some(title) = &opt.title {
             panel.set_title(title);
-        }
-
-        if let Some(parent) = &opt.parent {
-            panel.set_parent(parent);
-        }
-
-        let can_create_directories = opt.can_create_directories.unwrap_or(true);
-        panel.set_can_create_directories(match can_create_directories {
-            true => YES,
-            false => NO,
-        });
-
-        panel.set_can_choose_directories(YES);
-        panel.set_can_choose_files(NO);
-
-        panel
-    }
-
-    pub fn build_pick_folders(opt: &FileDialog) -> Self {
-        let mut panel = Panel::open_panel();
-
-        if let Some(path) = &opt.starting_directory {
-            panel.set_path(path, opt.file_name.as_deref());
-        }
-
-        if let Some(title) = &opt.title {
-            panel.set_title(title);
-        }
-
-        if let Some(parent) = &opt.parent {
-            panel.set_parent(parent);
         }
 
         let can = opt.can_create_directories.unwrap_or(true);
-        panel.set_can_create_directories(match can {
-            true => YES,
-            false => NO,
-        });
+        panel.set_can_create_directories(can);
 
-        panel.set_can_choose_directories(YES);
-        panel.set_can_choose_files(NO);
-        panel.set_allows_multiple_selection(YES);
+        unsafe { panel.setCanChooseDirectories(true) };
+        unsafe { panel.setCanChooseFiles(false) };
 
-        panel
+        Self::new(Id::into_super(panel), opt.parent.as_ref())
     }
 
-    pub fn build_pick_files(opt: &FileDialog) -> Self {
-        let mut panel = Panel::open_panel();
+    pub fn build_pick_folders(opt: &FileDialog, mtm: MainThreadMarker) -> Self {
+        let panel = unsafe { NSOpenPanel::openPanel(mtm) };
+
+        if let Some(path) = &opt.starting_directory {
+            panel.set_path(path, opt.file_name.as_deref());
+        }
+
+        if let Some(title) = &opt.title {
+            panel.set_title(title);
+        }
+
+        let can = opt.can_create_directories.unwrap_or(true);
+        panel.set_can_create_directories(can);
+
+        unsafe { panel.setCanChooseDirectories(true) };
+        unsafe { panel.setCanChooseFiles(false) };
+        unsafe { panel.setAllowsMultipleSelection(true) };
+
+        Self::new(Id::into_super(panel), opt.parent.as_ref())
+    }
+
+    pub fn build_pick_files(opt: &FileDialog, mtm: MainThreadMarker) -> Self {
+        let panel = unsafe { NSOpenPanel::openPanel(mtm) };
 
         if !opt.filters.is_empty() {
             panel.add_filters(&opt);
@@ -309,21 +282,14 @@ impl Panel {
             panel.set_title(title);
         }
 
-        if let Some(parent) = &opt.parent {
-            panel.set_parent(parent);
-        }
-
         if let Some(can) = opt.can_create_directories {
-            panel.set_can_create_directories(match can {
-                true => YES,
-                false => NO,
-            });
+            panel.set_can_create_directories(can);
         }
 
-        panel.set_can_choose_directories(NO);
-        panel.set_can_choose_files(YES);
-        panel.set_allows_multiple_selection(YES);
+        unsafe { panel.setCanChooseDirectories(false) };
+        unsafe { panel.setCanChooseFiles(true) };
+        unsafe { panel.setAllowsMultipleSelection(true) };
 
-        panel
+        Self::new(Id::into_super(panel), opt.parent.as_ref())
     }
 }